refactor: move interface types to dir
