+++ server/db/knex.ts
@@ -1,5 +1,5 @@
 import { knex as initKnex } from 'knex'
-import { Global } from 'interfaces/global'
+import { Global } from 'interfaces/Global'
 
 import knexConfig from './knexConfig'
