+++ server/interfaces/Attribute.ts
@@ -1,0 +1,4 @@
+export default interface Attribute {
+  key: string
+  value: string | number
+}
