+++ server/interfaces/Global.ts
