+++ server/interfaces/TokenMessage.ts
@@ -1,0 +1,6 @@
+export type TokenMessage = {
+  entries: {
+    capture_id: string
+  }[]
+  wallet_name: string
+}
