+++ server/models/rawCaptureFeature.ts
@@ -1,18 +1,5 @@
 import knex from 'db/knex'
-
-import { Attribute } from './captureFeature'
-
-export type RawCaptureFeature = {
-  id: number | string
-  lat: number
-  lon: number
-  field_user_id: number | string
-  field_username: string
-  attributes: Attribute[]
-  device_identifier: string | number
-  created_at: string
-  updated_at: string
-}
+import { RawCaptureFeature } from 'interfaces/RawCaptureFeature'
 
 export async function addRawCapture(rawCaptureFeature: RawCaptureFeature) {
   const wellKnownText = `POINT(${rawCaptureFeature.lon} ${rawCaptureFeature.lat})`
